# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão Escolar",
        "description": "API para gerenciamento de sistema escolar.",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/discipline": {
            "get": {
                "tags": [
                    "Discipline"
                ],
                "summary": "List Discipline with pagination",
                "description": "Get paginated list of Disciplines",
                "operationId": "f9e64f2cef5fc74312edd7855b6e70c0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for filtering discipline",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_column",
                        "in": "query",
                        "description": "Column to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Direction for sorting (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of discipline",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DisciplineResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Discipline"
                ],
                "summary": "Create a new Discipline",
                "description": "Create a new Discipline",
                "operationId": "f479d86e8eb7c69bf159aa22476c4c78",
                "requestBody": {
                    "description": "JSON object containing new Discipline data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DisciplineStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Discipline created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DisciplineResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/discipline/{id}": {
            "get": {
                "tags": [
                    "Discipline"
                ],
                "summary": "Retrieve details of a Discipline by ID",
                "description": "Show details of a specific Discipline",
                "operationId": "73f31bb24c004159b09f6abb0e0b37ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discipline ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discipline data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DisciplineResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Discipline not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Discipline"
                ],
                "summary": "Update an existing Discipline",
                "description": "Update an existing Discipline",
                "operationId": "a80635bff57c8dcfb3fcb98d3e9249ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discipline ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated Discipline data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DisciplineUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discipline updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DisciplineResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Discipline not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Discipline"
                ],
                "summary": "Delete a Discipline by ID",
                "description": "Delete a specific Discipline",
                "operationId": "8b5de322372cf71684eb1034c23bdb4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discipline ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Discipline deleted successfully"
                    },
                    "404": {
                        "description": "Discipline not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DisciplineStoreRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DisciplineUpdateRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "DisciplineResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}